move three scene logic to its own function/file, import and use in 3dModelViewer
